requirements.txt
streamlit>=1.33
python-docx>=0.8.11
chromadb>=0.5.4
sentence-transformers>=2.2.2
pypdf>=4.2.0
beautifulsoup4>=4.12.3
requests>=2.31.0
groq>=0.13.0
google-generativeai>=0.7.2

